+++ docs_src/encoders/pydantic.py
@@ -9,9 +9,17 @@ from lilya.encoders import Encoder, register_encoder
 
 class PydanticEncoder(Encoder):
     __type__ = BaseModel
+    # optional a name can be provided, so same named encoders are removed
+    name = "ModelDumpEncoder"
+
+    # is_type and is_type_structure are provided by Encoder.
+    # checked is the type provided by __type__.
 
     def serialize(self, obj: BaseModel) -> dict[str, Any]:
         return obj.model_dump()
+
+    def encode(self, structure: type[BaseModel], value: Any) -> Any:
+        return structure(**value)
 
 
 # A normal way