+++ tests/test_make_response.py
@@ -1,29 +1,31 @@
 from collections import deque
 from dataclasses import dataclass
 
+import orjson
 import pytest
 from attrs import define
 from msgspec import Struct
 from pydantic import BaseModel
 
+from lilya.encoders import Encoder, apply_structure
 from lilya.responses import make_response
 from lilya.routing import Path
 from lilya.testclient import create_client
-from tests.conftest import dont_run
 
 
-@dataclass
-class ResponseData:
-    name: str
-    age: int
+@pytest.mark.parametrize(
+    "json_encode_kwargs", [{}, {"json_encode_fn": orjson.dumps, "post_transform_fn": orjson.loads}]
+)
+def test_response_dataclass(json_encode_kwargs):
+    @dataclass
+    class ResponseData:
+        name: str
+        age: int
 
+    def home() -> ResponseData:
+        data = ResponseData(name="lilya", age=24)
+        return make_response(data, status_code=201, json_encode_extra_kwargs=json_encode_kwargs)
 
-def home() -> ResponseData:
-    data = ResponseData(name="lilya", age=24)
-    return make_response(data, status_code=201)
-
-
-def test_response_dataclass():
     with create_client(routes=[Path("/", home)]) as client:
         response = client.get("/")
 
@@ -41,7 +43,6 @@ def base_model() -> User:
     return make_response(data, status_code=208)
 
 
-@pytest.mark.skipif(dont_run, reason="Python 3.8 internals")
 def test_pydantic_custom_make_response():
     with create_client(routes=[Path("/", base_model)]) as client:
         response = client.get("/")
@@ -55,7 +56,6 @@ def base_model_list() -> User:
     return make_response(data)
 
 
-@pytest.mark.skipif(dont_run, reason="Python 3.8 internals")
 def test_pydantic_custom_make_response_list():
     with create_client(routes=[Path("/", base_model_list)]) as client:
         response = client.get("/")
@@ -127,10 +127,13 @@ def test_attrs_custom_make_response_list():
         assert response.json() == [{"name": "lilya", "age": 24}]
 
 
+@pytest.mark.parametrize(
+    "json_encode_kwargs", [{}, {"json_encode_fn": orjson.dumps, "post_transform_fn": orjson.loads}]
+)
 @pytest.mark.parametrize("value", ["1", 2, 2.2, None], ids=["str", "int", "float", "none"])
-def test_primitive_responses(value):
+def test_primitive_responses(value, json_encode_kwargs):
     def home():
-        return make_response(value, status_code=201)
+        return make_response(value, status_code=201, json_encode_extra_kwargs=json_encode_kwargs)
 
     with create_client(routes=[Path("/", home)]) as client:
         response = client.get("/")
@@ -151,6 +154,9 @@ def test_dict_response():
 
 
 @pytest.mark.parametrize(
+    "json_encode_kwargs", [{}, {"json_encode_fn": orjson.dumps, "post_transform_fn": orjson.loads}]
+)
+@pytest.mark.parametrize(
     "value",
     [
         [1, 2, 3, 4],
@@ -161,12 +167,88 @@ def test_dict_response():
     ],
     ids=["list", "set", "frozenset", "tuple", "deque"],
 )
-def test_structure_responses(value):
+def test_structure_responses(value, json_encode_kwargs):
     def home():
-        return make_response(value, status_code=201)
+        return make_response(value, status_code=201, json_encode_extra_kwargs=json_encode_kwargs)
 
     with create_client(routes=[Path("/", home)]) as client:
         response = client.get("/")
 
         assert response.status_code == 201
         assert response.json() == [1, 2, 3, 4]
+
+
+def test_custom_encoder_response():
+    class Foo:
+        a: int
+
+        def __init__(self, a: int):
+            self.a = a
+
+        def __eq__(self, other) -> bool:
+            return isinstance(other, type(self)) and other.a == self.a
+
+    @dataclass
+    class FooDataclass:
+        a: int
+
+    class FooEncoder(Encoder):
+        __type__ = Foo
+
+        def serialize(self, value: Foo) -> dict:
+            return {"a": value.a}
+
+        def encode(self, structure, value):
+            foo = Foo(value["a"])
+            return foo
+
+    def home():
+        return make_response(Foo(a=3), status_code=201, encoders=[FooEncoder()])
+
+    with create_client(routes=[Path("/", home)]) as client:
+        response = client.get("/")
+
+        assert response.status_code == 201
+        assert response.json() == {"a": 3}
+        result = apply_structure(Foo, response.json(), with_encoders=[FooEncoder()])
+        assert result == Foo(3)
+        assert result != FooDataclass(3)
+        # try again with defaults and a similar structured dataclass
+        assert apply_structure(FooDataclass, response.json()) == FooDataclass(3)
+
+
+@pytest.mark.parametrize(
+    "json_encode_kwargs", [{}, {"json_encode_fn": orjson.dumps, "post_transform_fn": orjson.loads}]
+)
+def test_custom_molding_only_encoder_response(json_encode_kwargs):
+    class Foo:
+        a: int
+
+        def __init__(self, a: int):
+            self.a = a
+
+        def __eq__(self, other) -> bool:
+            return isinstance(other, type(self)) and other.a == self.a
+
+    class FooEncoder(Encoder):
+        __type__ = Foo
+
+        def encode(self, structure, value):
+            foo = Foo(value["a"])
+            return foo
+
+    def home():
+        return make_response(
+            {"a": 3},
+            status_code=201,
+            encoders=[FooEncoder()],
+            json_encode_extra_kwargs=json_encode_kwargs,
+        )
+
+    with create_client(routes=[Path("/", home)]) as client:
+        response = client.get("/")
+
+        assert response.status_code == 201
+        assert response.json() == {"a": 3}
+        result = apply_structure(Foo, response.json(), with_encoders=[FooEncoder()])
+        assert result == Foo(3)