+++ docs/en/docs/release-notes.md
@@ -7,10 +7,20 @@ hide:
 
 ## Unreleased
 
+### Changed
+
+- Enhanced encoders:
+  - `ENCODER_TYPES` is now a context-variable.
+  - New encoders for datetime and date.
+  - Former `json_encoder` is now exported via public API as `json_encode`. The old name is still available in `_internal`.
+  - New method `apply_structure`, which allows input parsing.
+  - `make_response` does not support encoders as class anymore only as instance.
+
 ### Fixed
 
 - `typing_extensions` errors on python >= 3.10.
 - Address deprecation warnings in tests.
+- Fix encoders applied after simplifying response.
 
 ## 0.10.2